Heridores
Añadida capa para manejar heridores.
Modificada capa 'púas' para ahora llamarse 'trampas' y manejar objetos peligrosos con comportamientos.
Añadido objeto prefab Cactus.
Configurado comportamiento para Cactus, permitiendo ser un peligro estático, heridor.
Modificado comportamiento de púas para funcionar heredados desde heridores, permitiendo ahorrar código y simplificar comportamientos.
