+++ Assets/Prefabs/Interactuables/Peligros/Cactus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9be2ef7a64c43624e92191c5fda96519
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
