+++ Assets/Prefabs/Interactuables/Peligros/Puas Trampa.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90bb03b1961b4b94aadf1496b190921e
+guid: 18d1e163a7feea14abc6f6931c687ca9
 PrefabImporter:
   externalObjects: {}
   userData:
