+++ Assets/Scripts/Interactuables/Heridor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8cdfd26ac166ae740abd06d036a4624b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
