+++ ProjectSettings/TagManager.asset
@@ -25,7 +25,7 @@ TagManager:
   - Recolectable
   - Resorte
   - Llave
-  - Puas
+  - Trampa
   - Antorcha
   - Puerta
   - Bandera
@@ -35,7 +35,7 @@ TagManager:
   - Joya
   - Regenerador
   - Palanca
-  - 
+  - Heridor
   - 
   - 
   -
